--- .github/workflows/pixi_build_mac.yaml
@@ -1,52 +1,0 @@
-name: Build and test pixi-build
-# Builds FANS for macOS and Ubuntu and runs the tests.
-on:
-  push:
-    branches:
-      - main
-      - develop
-  pull_request:
-  workflow_dispatch:
-concurrency:
-  group: ${{ github.event_name }}-${{ github.workflow }}-${{ github.ref }}
-  cancel-in-progress: ${{github.event_name == 'pull_request'}}
-jobs:
-  build:
-    name: ${{ matrix.os }}
-    runs-on: ${{ matrix.os }}
-    env:
-      FANS_BUILD_DIR: build
-    strategy:
-      fail-fast: false
-      matrix:
-        os: [macos-13, macos-15, ubuntu-latest]
-    steps:
-      - name: Checkout code
-        uses: actions/checkout@v5
-
-      - name: Set up pixi
-        uses: prefix-dev/setup-pixi@v0.9.0
-
-      - uses: actions/upload-artifact@v4
-        if: failure()
-        with:
-          name: Pixi build (${{ matrix.os }})
-          path: .pixi/build
-      # ────────────────────────────────────────────────────────────────
-      # Pytest checks
-      # ────────────────────────────────────────────────────────────────
-      - name: Install Pixi Python deps
-        run: |
-          pixi --version
-          pixi install
-
-      - name: Run pytest checks on HDF5 output
-        run: pixi run test
-
-      - uses: actions/upload-artifact@v4
-        if: failure()
-        with:
-          name: ${{ matrix.os }} PyTest logs
-          path: |
-            **/pytest*.xml
-            **/.pytest_cache
